backend/requirements.txt
absl-py==2.2.2
aiohappyeyeballs==2.6.1
aiohttp==3.11.16
aiosignal==1.3.2
altair==5.5.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyascii==0.3.2
anyio==4.12.1
appnope==0.1.4
argon2-cffi==25.1.0
argon2-cffi-bindings==25.1.0
arrow==1.4.0
asttokens==3.0.0
async-lru==2.1.0
async-timeout==4.0.3
attrs==25.3.0
audioread==3.0.1
babel==2.17.0
backboard-sdk==1.4.6
backports.asyncio.runner==1.2.0
backports.tarfile==1.2.0
bangla==0.0.2
beautifulsoup4==4.14.3
bleach==6.3.0
blinker==1.9.0
blis==0.7.11
bnnumerizer==0.0.2
bnunicodenormalizer==0.1.7
build==1.2.2.post1
CacheControl==0.14.3
cachetools==6.2.4
catalogue==2.0.10
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
cleo==2.1.0
click==8.3.1
cloudpathlib==0.21.0
colorama==0.4.6
comm==0.2.3
confection==0.1.5
contourpy==1.2.1
coqpit==0.0.17
coqpit-config==0.1.2
coqui-tts==0.25.2
coqui-tts-trainer==0.2.2
coverage==7.13.1
crashtest==0.4.1
cryptography==46.0.3
csvw==3.5.1
cycler==0.12.1
cymem==2.0.11
Cython==3.0.12
dataclasses-json==0.6.7
dateparser==1.1.8
debugpy==1.8.15
decorator==5.2.1
defusedxml==0.7.1
distlib==0.3.9
distro==1.9.0
dlinfo==2.0.0
dnspython==2.8.0
docopt==0.6.2
dulwich==0.22.8
einops==0.8.1
email-validator==2.3.0
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.7.1/en_core_web_sm-3.7.1-py3-none-any.whl#sha256=86cc141f63942d4b2c5fcee06630fd6f904788d2f0ab005cce45aadb8fb73889
encodec==0.1.1
exceptiongroup==1.3.0
executing==2.2.0
fastapi==0.128.0
fastapi-cli==0.0.20
fastapi-cloud-cli==0.9.0
fastar==0.8.0
fastjsonschema==2.21.1
filelock==3.18.0
findpython==0.6.3
Flask==3.1.0
fonttools==4.57.0
fqdn==1.5.1
frozenlist==1.5.0
fsspec==2025.3.2
g2pkk==0.1.2
gitdb==4.0.12
GitPython==3.1.45
google==3.0.0
google-api-core==2.29.0
google-api-python-client==2.188.0
google-auth==2.41.1
google-auth-httplib2==0.3.0
google-auth-oauthlib==1.2.3
googleapis-common-protos==1.72.0
gruut==2.4.0
gruut-ipa==0.13.0
gruut_lang_de==2.0.1
gruut_lang_en==2.0.1
gruut_lang_es==2.0.1
gruut_lang_fr==2.0.2
h11==0.16.0
hangul-romanize==0.1.0
httpcore==1.0.9
httplib2==0.31.1
httptools==0.7.1
httpx==0.28.1
httpx-sse==0.4.1
huggingface-hub==0.30.1
idna==3.11
importlib_metadata==8.7.0
inflect==7.5.0
iniconfig==2.3.0
installer==0.7.0
ipykernel==7.2.0
ipython==8.37.0
ipywidgets==8.1.8
isodate==0.7.2
isoduration==20.11.0
itsdangerous==2.2.0
jamo==0.4.1
jaraco.classes==3.4.0
jaraco.context==6.0.1
jaraco.functools==4.1.0
jedi==0.19.2
jieba==0.42.1
Jinja2==3.1.6
jiter==0.10.0
joblib==1.4.2
json5==0.13.0
jsonlines==1.2.0
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter==1.1.1
jupyter-console==6.6.3
jupyter-events==0.12.0
jupyter-lsp==2.3.0
jupyter_client==8.8.0
jupyter_core==5.8.1
jupyter_server==2.17.0
jupyter_server_terminals==0.5.4
jupyterlab==4.5.4
jupyterlab_pygments==0.3.0
jupyterlab_server==2.28.0
jupyterlab_widgets==3.0.16
keyring==25.6.0
kiwisolver==1.4.8
langchain==0.3.26
langchain-community==0.3.27
langchain-core==0.3.70
langchain-openai==0.3.28
langchain-text-splitters==0.3.8
langcodes==3.5.0
langgraph==0.6.3
langgraph-checkpoint==2.1.1
langgraph-prebuilt==0.6.3
langgraph-sdk==0.2.0
langsmith==0.4.8
language-tags==1.2.0
language_data==1.3.0
lazy_loader==0.4
librosa==0.11.0
llvmlite==0.43.0
marisa-trie==1.2.1
Markdown==3.7
markdown-it-py==4.0.0
MarkupSafe==3.0.3
marshmallow==3.26.1
matplotlib==3.8.4
matplotlib-inline==0.1.7
mdurl==0.1.2
mistune==3.2.0
monotonic-alignment-search==0.1.1
more-itertools==10.6.0
mpmath==1.3.0
msgpack==1.1.0
multidict==6.2.0
murmurhash==1.0.12
mutagen==1.47.0
mypy_extensions==1.1.0
narwhals==2.0.1
nbclient==0.10.4
nbconvert==7.17.0
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==2.8.8
nltk==3.9.1
notebook==7.5.3
notebook_shim==0.2.4
num2words==0.5.14
numba==0.60.0
numpy==1.26.4
oauthlib==3.3.1
openai==1.98.0
orjson==3.11.0
ormsgpack==1.10.0
overrides==7.7.0
packaging==25.0
pandas==1.5.3
pandocfilters==1.5.1
parso==0.8.4
pbs-installer==2025.5.17
pdfminer.six==20250506
pdfplumber==0.11.7
pexpect==4.9.0
phonemizer==3.3.0
phonenumbers==9.0.10
pillow==11.1.0
pkginfo==1.12.1.2
platformdirs==4.3.7
pluggy==1.6.0
poetry==2.1.3
poetry-core==2.1.3
pooch==1.8.2
preshed==3.0.9
prometheus_client==0.24.1
prompt_toolkit==3.0.51
propcache==0.3.1
proto-plus==1.27.0
protobuf==6.33.4
psutil==7.0.0
psycopg2-binary==2.9.10
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==21.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycountry==22.3.5
pycparser==2.23
pydantic==2.12.5
pydantic-extra-types==2.11.0
pydantic-settings==2.12.0
pydantic_core==2.41.5
pydeck==0.9.1
Pygments==2.19.2
pynndescent==0.5.13
pyparsing==3.3.1
pypdfium2==4.30.0
pypinyin==0.54.0
pyproject_hooks==1.2.0
pysbd==0.3.4
pytest==9.0.2
pytest-asyncio==1.3.0
pytest-cov==7.0.0
python-crfsuite==0.9.11
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-json-logger==4.0.0
python-multipart==0.0.21
python-telegram-bot==22.5
pytz==2025.2
PyYAML==6.0.3
pyzmq==27.0.0
RapidFuzz==3.13.0
rdflib==7.1.4
referencing==0.36.2
regex==2024.11.6
requests==2.32.5
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rfc3339-validator==0.1.4
rfc3986==1.5.0
rfc3986-validator==0.1.1
rich==14.2.0
rich-toolkit==0.17.1
rignore==0.7.6
rpds-py==0.24.0
rsa==4.9.1
safetensors==0.5.3
scikit-learn==1.6.1
scipy==1.15.2
segments==2.3.0
Send2Trash==2.1.0
sentry-sdk==2.49.0
shellingham==1.5.4
six==1.17.0
smart-open==7.1.0
smmap==5.0.2
sniffio==1.3.1
soundfile==0.13.1
soupsieve==2.8.1
soxr==0.5.0.post1
spacy==3.7.5
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SQLAlchemy==2.0.41
srsly==2.5.1
stack-data==0.6.3
starlette==0.50.0
streamlit==1.48.0
SudachiDict-core==20250129
SudachiPy==0.6.10
sympy==1.13.1
tenacity==9.1.2
tensorboard==2.19.0
tensorboard-data-server==0.7.2
termcolor==2.3.0
terminado==0.18.1
thinc==8.2.5
threadpoolctl==3.6.0
tiktoken==0.9.0
tinycss2==1.4.0
tokenizers==0.20.3
toml==0.10.2
tomli==2.2.1
tomlkit==0.13.2
torch==2.6.0
torchaudio==2.6.0
torchvision==0.21.0
tornado==6.5.1
toutatis==1.31
tqdm==4.67.1
trainer==0.0.36
traitlets==5.14.3
transformers==4.46.2
trove-classifiers==2025.5.9.12
typeguard==4.4.2
typer==0.21.1
typing-inspect==0.9.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
tzlocal==5.3.1
umap-learn==0.5.7
Unidecode==1.3.8
uri-template==1.3.0
uritemplate==4.2.0
urllib3==2.6.3
uvicorn==0.40.0
uvloop==0.22.1
virtualenv==20.31.2
wasabi==1.1.3
watchfiles==1.1.1
wcwidth==0.2.13
weasel==0.4.1
webcolors==25.10.0
webencodings==0.5.1
websocket-client==1.9.0
websockets==16.0
Werkzeug==3.1.3
widgetsnbextension==4.0.15
wrapt==1.17.2
xattr==1.1.4
xxhash==3.5.0
yarl==1.18.3
zipp==3.21.0
zstandard==0.23.0
